requirements.txt
addict==2.4.0
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.5.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
bayesian-optimization==1.4.3
bleach==6.0.0
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==2.1.0
clearml==1.12.2
clearml-agent==1.6.1
click==8.1.3
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.10.0
data==0.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.1.1
djangorestframework==3.14.0
docopt==0.6.2
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.108.0
filelock==3.12.2
flake8==6.0.0
Flask==2.2.2
fonttools==4.42.1
frozenlist==1.4.1
funcsigs==1.0.2
furl==2.1.3
future==1.0.0
grip==4.6.1
h11==0.14.0
httpie==3.2.2
httptools==0.6.1
idna==2.10
imageio==2.31.6
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kaggle==1.5.16
kiwisolver==1.4.5
kornia==0.7.0
latex==0.7.0
-e git+https://github.com/cvg/LightGlue.git@5a9e87d85905a6770547b86a79fa0ae227d59f36#egg=lightglue
magic-filter==1.0.12
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mlxtend==0.23.1
mmcv-full==1.3.15
mmengine==0.9.0
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.4
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
orderedmultidict==1.0.1
packaging==23.1
pandas==2.2.0
parso==0.8.3
path-and-address==2.0.1
pathlib==1.0.1
pathlib2==2.3.7.post1
patsy==0.5.6
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyarrow==15.0.0
pycodestyle==2.10.0
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.0.1
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytube==15.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.28.1
requests-toolbelt==1.0.0
rich==13.6.0
roboflow==1.1.7
rpds-py==0.10.2
scikit-learn==1.4.1.post1
scipy==1.11.2
seaborn==0.13.2
shutilwhich==1.1.0
six==1.16.0
sklearn==0.0.post11
sniffio==1.3.0
sqlparse==0.4.2
stack-data==0.6.3
starlette==0.32.0.post1
statsmodels==0.14.1
supervision==0.16.0
sympy==1.12
tempdir==0.7.1
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.2.0
tk==0.1.0
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.11
uvicorn==0.25.0
virtualenv==20.24.5
watchfiles==0.21.0
wcwidth==0.2.8
webencodings==0.5.1
websockets==12.0
Werkzeug==2.2.2
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
